Organize the code to have the devices defined in independents files
