+++ main.c
@@ -33,2193 +33,19 @@
 #include <libevdev-1.0/libevdev/libevdev.h>
 #include <libevdev-1.0/libevdev/libevdev-uinput.h>
 
+#include "main.h"
 #include "parser.h"
+#include "devices/rg552.h"
+#include "devices/chi.h"
+#include "devices/ogs.h"
+#include "devices/ogx.h"
+#include "devices/rk2020.h"
+#include "devices/anbernic.h"
 
 #ifdef TESTING_LAG
 #include <sys/time.h>
 #include <stdbool.h>
 #endif
-
-typedef void (*function_type)(int, int, int);
-static function_type handleEventFunction;
-
-struct libevdev* dev_joypad;
-int fd_ev_joypad;
-int rc_joypad;
-struct input_event ev_joypad;
-
-static int uinp_fd = -1;
-struct uinput_user_dev uidev;
-
-int debug = 0;
-char quit_command[100];
-
-// Define variables to be used for key codes
-int back_key;
-int start_key;
-int a_key;
-int b_key;
-int x_key;
-int y_key;
-int up_key;
-int down_key;
-int left_key;
-int right_key;
-int l1_key;
-int l2_key;
-int l3_key;
-int r1_key;
-int r2_key;
-int r3_key;
-int deadzone_y;
-int deadzone_x;
-
-short back = KEY_ESC;
-short start = KEY_ENTER;
-short a = KEY_V;
-short b = KEY_C;
-short x = KEY_X;
-short y = KEY_A;
-short l1 = KEY_B;
-short l2 = KEY_R;
-short l3 = BTN_LEFT;
-short r1 = KEY_D;
-short r2 = KEY_S;
-short r3 = BTN_RIGHT;
-short up = KEY_UP;
-short down = KEY_DOWN;
-short left = KEY_LEFT;
-short right = KEY_RIGHT;
-short left_analog_up = KEY_UP;
-short left_analog_down = KEY_DOWN;
-short left_analog_left = KEY_LEFT;
-short left_analog_right = KEY_RIGHT;
-short right_analog_up = KEY_UP;
-short right_analog_down = KEY_DOWN;
-short right_analog_left = KEY_LEFT;
-short right_analog_right = KEY_RIGHT;
-
-int left_analog_x = 0;
-int left_analog_y = 0;
-int right_analog_x = 0;
-int right_analog_y = 0;
-
-int left_analog_mouse = 0;
-int right_analog_mouse = 0;
-
-int hold = 0;
-
-void emit(int type, int code, int val) {
-  struct input_event ev;
-
-  ev.type = type;
-  ev.code = code;
-  ev.value = val;
-  ev.time.tv_sec = 0;
-  ev.time.tv_usec = 0;
-
-  write(uinp_fd, &ev, sizeof(ev));
-}
-
-// ******************* RG552
-
-void handle_event_rg552(int type, int code, int value) {
-  if (type == 1) {
-    if (code == back_key && value == 1) {
-      emit(EV_KEY, back, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == back_key && value == 0) {
-      emit(EV_KEY, back, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == start_key && value == 1) {
-      if (hold == 1) {
-        system(quit_command);
-        system("sleep 0.2");
-        exit (0);
-      }
-      emit(EV_KEY, start, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == start_key && value == 0) {
-      emit(EV_KEY, start, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == a_key && value == 1) {
-      emit(EV_KEY, a, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == a_key && value == 0) {
-      emit(EV_KEY, a, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == b_key && value == 1) {
-      emit(EV_KEY, b, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == b_key && value == 0) {
-      emit(EV_KEY, b, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == x_key && value == 1) {
-      emit(EV_KEY, x, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == x_key && value == 0) {
-      emit(EV_KEY, x, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == y_key && value == 1) {
-      emit(EV_KEY, y, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == y_key && value == 0) {
-      emit(EV_KEY, y, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == l1_key && value == 1) {
-      emit(EV_KEY, l1, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == l1_key && value == 0) {
-      emit(EV_KEY, l1, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == l2_key && value == 1) {
-      emit(EV_KEY, l2, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == l2_key && value == 0) {
-      emit(EV_KEY, l2, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == l3_key && value == 1) {
-      hold = 1;
-      emit(EV_KEY, l3, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == l3_key && value == 0) {
-      hold = 0;
-      emit(EV_KEY, l3, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == r1_key && value == 1) {
-      emit(EV_KEY, r1, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == r1_key && value == 0) {
-      emit(EV_KEY, r1, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == r2_key && value == 1) {
-      emit(EV_KEY, r2, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == r2_key && value == 0) {
-      emit(EV_KEY, r2, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == r3_key && value == 1) {
-      hold = 1;
-      emit(EV_KEY, r3, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == r3_key && value == 0) {
-      hold = 0;
-      emit(EV_KEY, r3, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    //dpad
-    if (code == up_key && value == 1) {
-      emit(EV_KEY, up, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == up_key && value == 0) {
-      emit(EV_KEY, up, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == down_key && value == 1) {
-      emit(EV_KEY, down, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == down_key && value == 0) {
-      emit(EV_KEY, down, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == left_key && value == 1) {
-      emit(EV_KEY, left, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == left_key && value == 0) {
-      emit(EV_KEY, left, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == right_key && value == 1) {
-      emit(EV_KEY, right, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == right_key && value == 0) {
-      emit(EV_KEY, right, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-  }
-
-  // analog
-  if (type == 3) {
-    // mouse movement, right analog
-    if (right_analog_mouse) {
-      if (code == 4) { // up/down
-        if (value > deadzone_y) {
-          emit(EV_REL, REL_Y, (value/64));
-          emit(EV_SYN, SYN_REPORT, 0);
-        }
-
-        if (value < deadzone_x) {
-          emit(EV_REL, REL_Y, (value/64));
-          emit(EV_SYN, SYN_REPORT, 0);
-        }
-      }
-      else if (code == 3) { // left/right
-        if (value > deadzone_y) {
-          emit(EV_REL, REL_X, (value/64));
-          emit(EV_SYN, SYN_REPORT, 0);
-        }
-
-        if (value < deadzone_x) {
-          emit(EV_REL, REL_X, (value/64));
-          emit(EV_SYN, SYN_REPORT, 0);
-        }
-      }
-    }
-    else {
-      if (code == 4) { // up/down
-        if (value > deadzone_y) {
-          emit(EV_KEY, right_analog_down, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          right_analog_y = 2;
-        }
-        else {
-          if (right_analog_y > 0 || right_analog_x > 0) {
-            emit(EV_KEY, right_analog_down, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            right_analog_y = right_analog_y - 1;
-          }
-        }
-        if (value < deadzone_x) {
-          emit(EV_KEY, right_analog_up, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          right_analog_y = 2;
-        }
-        else {
-          if (right_analog_y > 0 || right_analog_x > 0) {
-            emit(EV_KEY, right_analog_up, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            right_analog_y = right_analog_y - 1;
-          }
-        }
-      }
-
-      if (code == 3) { // left/right
-        if (value > deadzone_y) {
-          emit(EV_KEY, right_analog_right, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          right_analog_x = 2;
-        }
-        else {
-          if (right_analog_x > 0 || right_analog_y > 0) {
-            emit(EV_KEY, right_analog_right, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            right_analog_x = right_analog_x - 1;
-          }
-        }
-        if (value < deadzone_x) {
-          emit(EV_KEY, right_analog_left, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          right_analog_x = 2;
-        }
-        else {
-          if (right_analog_x > 0 || right_analog_y > 0) {
-            emit(EV_KEY, right_analog_left, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            right_analog_x = right_analog_x - 1;
-          }
-        }
-      }
-    }
-    // mouse movement, left analog
-    if (left_analog_mouse) {
-      if (code == 1) { // up/down
-        if (value > deadzone_y) {
-          emit(EV_REL, REL_Y, (value/64));
-          emit(EV_SYN, SYN_REPORT, 0);
-        }
-
-        if (value < deadzone_x) {
-          emit(EV_REL, REL_Y, (value/64));
-          emit(EV_SYN, SYN_REPORT, 0);
-        }
-      }
-      else if (code == 0) { // left/right
-        if (value > deadzone_y) {
-          emit(EV_REL, REL_X, (value/64));
-          emit(EV_SYN, SYN_REPORT, 0);
-        }
-
-        if (value < deadzone_x) {
-          emit(EV_REL, REL_X, (value/64));
-          emit(EV_SYN, SYN_REPORT, 0);
-        }
-      }
-    }
-    else {
-      if (code == 1) { // left analog y axis
-        if (value > deadzone_y) {
-          emit(EV_KEY, left_analog_down, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          left_analog_y = 2;
-        }
-        else {
-          if (left_analog_y > 0 || left_analog_x > 0) {
-            emit(EV_KEY, left_analog_down, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            left_analog_y = left_analog_y - 1;
-          }
-        }
-        if (value < deadzone_x) {
-          emit(EV_KEY, left_analog_up, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          left_analog_y = 2;
-        }
-        else {
-          if (left_analog_y > 0 || left_analog_x > 0) {
-            emit(EV_KEY, left_analog_up, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            left_analog_y = left_analog_y - 1;
-          }
-        }
-      }
-      if (code == 0) { // left analogue x axis
-        if (value > deadzone_y) {
-          emit(EV_KEY, left_analog_right, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          left_analog_x = 2;
-        }
-        else {
-          if (left_analog_x > 0 || left_analog_y > 0) {
-            emit(EV_KEY, left_analog_right, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            left_analog_x = left_analog_x - 1;
-          }
-        }
-        if (value < deadzone_x) {
-          emit(EV_KEY, left_analog_left, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          left_analog_x = 2;
-        }
-        else {
-          if (left_analog_x > 0 || left_analog_y > 0) {
-            emit(EV_KEY, left_analog_left, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            left_analog_x = left_analog_x - 1;
-          }
-        }
-      }
-    }
-  }
-}
-
-void config_rg552(char *inputstr)
-{
-#ifdef DEBUG
-  printf("OGA Contols - Configuring 'RG552' device\n");
-#endif
-  handleEventFunction = &handle_event_rg552;
-  back_key = 314;
-  start_key = 315;
-  a_key = 305;
-  b_key = 304;
-  x_key = 307;
-  y_key = 308;
-  up_key = 544;
-  down_key = 545;
-  left_key = 546;
-  right_key = 547;
-  l1_key = 310;
-  l2_key = 312;
-  l3_key = 317;
-  r1_key = 311;
-  r2_key = 313;
-  r3_key = 318;
-  deadzone_y = 300;
-  deadzone_x = -300;
-  strcpy(inputstr, "/dev/input/by-path/platform-singleadc-joypad-event-joystick");
-}
-
-// ******************* CHI
-
-void handle_event_chi(int type, int code, int value) {
-  if (type == 1) {
-    if (code == back_key && value == 1) {
-      emit(EV_KEY, back, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == back_key && value == 0) {
-      emit(EV_KEY, back, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == start_key && value == 1) {
-      if (hold == 1) {
-        system(quit_command);
-        system("sleep 0.2");
-        exit (0);
-      }
-      emit(EV_KEY, start, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == start_key && value == 0) {
-      emit(EV_KEY, start, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == a_key && value == 1) {
-      emit(EV_KEY, a, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == a_key && value == 0) {
-      emit(EV_KEY, a, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == b_key && value == 1) {
-      emit(EV_KEY, b, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == b_key && value == 0) {
-      emit(EV_KEY, b, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == x_key && value == 1) {
-      emit(EV_KEY, x, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == x_key && value == 0) {
-      emit(EV_KEY, x, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == y_key && value == 1) {
-      emit(EV_KEY, y, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == y_key && value == 0) {
-      emit(EV_KEY, y, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == l1_key && value == 1) {
-      emit(EV_KEY, l1, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == l1_key && value == 0) {
-      emit(EV_KEY, l1, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == l2_key && value == 1) {
-      emit(EV_KEY, l2, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == l2_key && value == 0) {
-      emit(EV_KEY, l2, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == l3_key && value == 1) {
-      hold = 1;
-      emit(EV_KEY, l3, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == l3_key && value == 0) {
-      hold = 0;
-      emit(EV_KEY, l3, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == r1_key && value == 1) {
-      emit(EV_KEY, r1, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == r1_key && value == 0) {
-      emit(EV_KEY, r1, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == r2_key && value == 1) {
-      emit(EV_KEY, r2, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == r2_key && value == 0) {
-      emit(EV_KEY, r2, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == r3_key && value == 1) {
-      hold = 1;
-      emit(EV_KEY, r3, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == r3_key && value == 0) {
-      hold = 0;
-      emit(EV_KEY, r3, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    //dpad
-    if (code == up_key && value == 1) {
-      emit(EV_KEY, up, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == up_key && value == 0) {
-      emit(EV_KEY, up, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == down_key && value == 1) {
-      emit(EV_KEY, down, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == down_key && value == 0) {
-      emit(EV_KEY, down, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == left_key && value == 1) {
-      emit(EV_KEY, left, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == left_key && value == 0) {
-      emit(EV_KEY, left, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == right_key && value == 1) {
-      emit(EV_KEY, right, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == right_key && value == 0) {
-      emit(EV_KEY, right, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-  }
-
-  // analog
-  if (type == 3) {
-    // mouse movement, right analog
-    if (right_analog_mouse) {
-      if (code == 3) { // up/down
-        if (value > deadzone_y) {
-          emit(EV_REL, REL_Y, (value/64));
-          emit(EV_SYN, SYN_REPORT, 0);
-        }
-
-        if (value < deadzone_x) {
-          emit(EV_REL, REL_Y, (value/64));
-          emit(EV_SYN, SYN_REPORT, 0);
-        }
-      }
-      else if (code == 4) { // left/right
-        if (value > deadzone_y) {
-          emit(EV_REL, REL_X, (value/64));
-          emit(EV_SYN, SYN_REPORT, 0);
-        }
-
-        if (value < deadzone_x) {
-          emit(EV_REL, REL_X, (value/64));
-          emit(EV_SYN, SYN_REPORT, 0);
-        }
-      }
-    }
-    else {
-      if (code == 3) { // up/down
-        if (value > deadzone_y) {
-          emit(EV_KEY, right_analog_down, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          right_analog_y = 2;
-        }
-        else {
-          if (right_analog_y > 0 || right_analog_x > 0) {
-            emit(EV_KEY, right_analog_down, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            right_analog_y = right_analog_y - 1;
-          }
-        }
-        if (value < deadzone_x) {
-          emit(EV_KEY, right_analog_up, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          right_analog_y = 2;
-        }
-        else {
-          if (right_analog_y > 0 || right_analog_x > 0) {
-            emit(EV_KEY, right_analog_up, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            right_analog_y = right_analog_y - 1;
-          }
-        }
-      }
-
-      if (code == 4) { // left/right
-        if (value > deadzone_y) {
-          emit(EV_KEY, right_analog_right, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          right_analog_x = 2;
-        }
-        else {
-          if (right_analog_x > 0 || right_analog_y > 0) {
-            emit(EV_KEY, right_analog_right, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            right_analog_x = right_analog_x - 1;
-          }
-        }
-        if (value < deadzone_x) {
-          emit(EV_KEY, right_analog_left, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          right_analog_x = 2;
-        }
-        else {
-          if (right_analog_x > 0 || right_analog_y > 0) {
-            emit(EV_KEY, right_analog_left, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            right_analog_x = right_analog_x - 1;
-          }
-        }
-      }
-    }
-    // mouse movement, left analog
-    if (left_analog_mouse) {
-      if (code == 0) { // up/down
-        if (value > deadzone_y) {
-          emit(EV_REL, REL_Y, (value/64));
-          emit(EV_SYN, SYN_REPORT, 0);
-        }
-
-        if (value < deadzone_x) {
-          emit(EV_REL, REL_Y, (value/64));
-          emit(EV_SYN, SYN_REPORT, 0);
-        }
-      }
-      else if (code == 1) { // left/right
-        if (value > deadzone_y) {
-          emit(EV_REL, REL_X, (value/64));
-          emit(EV_SYN, SYN_REPORT, 0);
-        }
-
-        if (value < deadzone_x) {
-          emit(EV_REL, REL_X, (value/64));
-          emit(EV_SYN, SYN_REPORT, 0);
-        }
-      }
-    }
-    else {
-      if (code == 0) { // left analog y axis
-        if (value > deadzone_y) {
-          emit(EV_KEY, left_analog_down, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          left_analog_y = 2;
-        }
-        else {
-          if (left_analog_y > 0 || left_analog_x > 0) {
-            emit(EV_KEY, left_analog_down, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            left_analog_y = left_analog_y - 1;
-          }
-        }
-        if (value < deadzone_x) {
-          emit(EV_KEY, left_analog_up, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          left_analog_y = 2;
-        }
-        else {
-          if (left_analog_y > 0 || left_analog_x > 0) {
-            emit(EV_KEY, left_analog_up, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            left_analog_y = left_analog_y - 1;
-          }
-        }
-      }
-      if (code == 1) { // left analogue x axis
-        if (value > deadzone_y) {
-          emit(EV_KEY, left_analog_right, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          left_analog_x = 2;
-        }
-        else {
-          if (left_analog_x > 0 || left_analog_y > 0) {
-            emit(EV_KEY, left_analog_right, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            left_analog_x = left_analog_x - 1;
-          }
-        }
-        if (value < deadzone_x) {
-          emit(EV_KEY, left_analog_left, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          left_analog_x = 2;
-        }
-        else {
-          if (left_analog_x > 0 || left_analog_y > 0) {
-            emit(EV_KEY, left_analog_left, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            left_analog_x = left_analog_x - 1;
-          }
-        }
-      }
-    }
-  }
-}
-
-void config_chi(char *inputstr)
-{
-#ifdef DEBUG
-  printf("OGA Contols - Configuring 'CHI' device\n");
-#endif
-  handleEventFunction = &handle_event_chi;
-  back_key = 314;
-  start_key = 315;
-  a_key = 305;
-  b_key = 304;
-  x_key = 307;
-  y_key = 308;
-  up_key = 544;
-  down_key = 545;
-  left_key = 546;
-  right_key = 547;
-  l1_key = 310;
-  l2_key = 312;
-  l3_key = 704; /*1 key*/
-  r1_key = 311;
-  r2_key = 313;
-  r3_key = 705; /*2 key*/
-  deadzone_y = 300;
-  deadzone_x = -300;
-  strcpy(inputstr, "/dev/input/by-path/platform-gameforce-gamepad-event-joystick");
-}
-
-// ******************* OGS
-
-void handle_event_ogs(int type, int code, int value) {
-  if (type == 1) {
-    if (code == back_key && value == 1) {
-      hold = 1;
-      emit(EV_KEY, back, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == back_key && value == 0) {
-      hold = 0;
-      emit(EV_KEY, back, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == start_key && value == 1) {
-      if (hold == 1) {
-        system(quit_command);
-        system("sleep 0.2");
-        exit (0);
-      }
-      emit(EV_KEY, start, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == start_key && value == 0) {
-      emit(EV_KEY, start, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == a_key && value == 1) {
-      emit(EV_KEY, a, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == a_key && value == 0) {
-      emit(EV_KEY, a, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == b_key && value == 1) {
-      emit(EV_KEY, b, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == b_key && value == 0) {
-      emit(EV_KEY, b, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == x_key && value == 1) {
-      emit(EV_KEY, x, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == x_key && value == 0) {
-      emit(EV_KEY, x, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == y_key && value == 1) {
-      emit(EV_KEY, y, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == y_key && value == 0) {
-      emit(EV_KEY, y, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == l1_key && value == 1) {
-      emit(EV_KEY, l1, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == l1_key && value == 0) {
-      emit(EV_KEY, l1, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == l2_key && value == 1) {
-      emit(EV_KEY, l2, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == l2_key && value == 0) {
-      emit(EV_KEY, l2, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == l3_key && value == 1) {
-      emit(EV_KEY, l3, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == l3_key && value == 0) {
-      emit(EV_KEY, l3, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == r1_key && value == 1) {
-      emit(EV_KEY, r1, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == r1_key && value == 0) {
-      emit(EV_KEY, r1, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == r2_key && value == 1) {
-      emit(EV_KEY, r2, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == r2_key && value == 0) {
-      emit(EV_KEY, r2, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == r3_key && value == 1) {
-      emit(EV_KEY, r3, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == r3_key && value == 0) {
-      emit(EV_KEY, r3, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    //dpad
-    if (code == up_key && value == 1) {
-      emit(EV_KEY, up, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == up_key && value == 0) {
-      emit(EV_KEY, up, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == down_key && value == 1) {
-      emit(EV_KEY, down, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == down_key && value == 0) {
-      emit(EV_KEY, down, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == left_key && value == 1) {
-      emit(EV_KEY, left, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == left_key && value == 0) {
-      emit(EV_KEY, left, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == right_key && value == 1) {
-      emit(EV_KEY, right, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == right_key && value == 0) {
-      emit(EV_KEY, right, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-  }
-
-  // analog
-  if (type == 3) {
-    // mouse movement, right analog
-    if (right_analog_mouse) {
-      if (code == 4) { // up/down
-        if (value > deadzone_y) {
-          emit(EV_REL, REL_Y, -1);
-          emit(EV_SYN, SYN_REPORT, 0);
-        }
-
-        if (value < deadzone_x) {
-          emit(EV_REL, REL_Y, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-        }
-      }
-      else if (code == 3) { // left/right
-        if (value > deadzone_y) {
-          emit(EV_REL, REL_X, -1);
-          emit(EV_SYN, SYN_REPORT, 0);
-        }
-
-        if (value < deadzone_x) {
-          emit(EV_REL, REL_X, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-        }
-      }
-    }
-    else {
-      if (code == 4) { // up/down
-        if (value > deadzone_y) {
-          emit(EV_KEY, right_analog_down, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          right_analog_y = 2;
-        }
-        else {
-          if (right_analog_y > 0 || right_analog_x > 0) {
-            emit(EV_KEY, right_analog_down, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            right_analog_y = right_analog_y - 1;
-          }
-        }
-        if (value < deadzone_x) {
-          emit(EV_KEY, right_analog_up, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          right_analog_y = 2;
-        }
-        else {
-          if (right_analog_y > 0 || right_analog_x > 0) {
-            emit(EV_KEY, right_analog_up, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            right_analog_y = right_analog_y - 1;
-          }
-        }
-      }
-
-      if (code == 3) { // left/right
-        if (value > deadzone_y) {
-          emit(EV_KEY, right_analog_right, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          right_analog_x = 2;
-        }
-        else {
-          if (right_analog_x > 0 || right_analog_y > 0) {
-            emit(EV_KEY, right_analog_right, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            right_analog_x = right_analog_x - 1;
-          }
-        }
-        if (value < deadzone_x) {
-          emit(EV_KEY, right_analog_left, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          right_analog_x = 2;
-        }
-        else {
-          if (right_analog_x > 0 || right_analog_y > 0) {
-            emit(EV_KEY, right_analog_left, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            right_analog_x = right_analog_x - 1;
-          }
-        }
-      }
-    }
-    // mouse movement, left analog
-    if (left_analog_mouse) {
-      if (code == 1) { // up/down
-        if (value > deadzone_y) {
-          emit(EV_REL, REL_Y, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          left_analog_y = 2;
-        }
-        else {
-          if (left_analog_y > 0 || left_analog_x > 0) {
-            emit(EV_REL, REL_Y, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            left_analog_y = left_analog_y - 1;
-          }
-        }
-        if (value < deadzone_x) {
-          emit(EV_REL, REL_Y, -1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          left_analog_y = 2;
-        }
-        else {
-          if (left_analog_y > 0 || left_analog_x > 0) {
-            emit(EV_REL, REL_Y, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            left_analog_y = left_analog_y - 1;
-          }
-        }
-      }
-      if (code == 0) { // left analog x axis
-        if (value > deadzone_y) {
-          emit(EV_REL, REL_X, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          left_analog_x = 2;
-        }
-        else {
-          if (left_analog_x > 0 || left_analog_y > 0) {
-            emit(EV_REL, REL_X, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            left_analog_x = left_analog_x - 1;
-          }
-        }
-        if (value < deadzone_x) {
-          emit(EV_REL, REL_X, -1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          left_analog_x = 2;
-        }
-        else {
-          if (left_analog_x > 0 || left_analog_y > 0) {
-            emit(EV_REL, REL_X, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            left_analog_x = left_analog_x - 1;
-          }
-        }
-      }
-    }
-    else {
-      if (code == 1) { // left analog y axis
-        if (value > deadzone_y) {
-          emit(EV_KEY, left_analog_down, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          left_analog_y = 2;
-        }
-        else {
-          if (left_analog_y > 0 || left_analog_x > 0) {
-            emit(EV_KEY, left_analog_down, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            left_analog_y = left_analog_y - 1;
-          }
-        }
-        if (value < deadzone_x) {
-          emit(EV_KEY, left_analog_up, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          left_analog_y = 2;
-        }
-        else {
-          if (left_analog_y > 0 || left_analog_x > 0) {
-            emit(EV_KEY, left_analog_up, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            left_analog_y = left_analog_y - 1;
-          }
-        }
-      }
-      if (code == 0) { // left analog x axis
-        if (value > deadzone_y) {
-          emit(EV_KEY, left_analog_right, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          left_analog_x = 2;
-        }
-        else {
-          if (left_analog_x > 0 || left_analog_y > 0) {
-            emit(EV_KEY, left_analog_right, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            left_analog_x = left_analog_x - 1;
-          }
-        }
-        if (value < deadzone_x) {
-          emit(EV_KEY, left_analog_left, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          left_analog_x = 2;
-        }
-        else {
-          if (left_analog_x > 0 || left_analog_y > 0) {
-            emit(EV_KEY, left_analog_left, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            left_analog_x = left_analog_x - 1;
-          }
-        }
-      }
-    }
-  }
-}
-
-void config_ogs(char *inputstr)
-{
-#ifdef DEBUG
-  printf("OGA Contols - Configuring 'OGS' device\n");
-#endif
-  handleEventFunction = &handle_event_ogs;
-  back_key = 704;
-  start_key = 705;
-  a_key = 305;
-  b_key = 304;
-  x_key = 307;
-  y_key = 308;
-  up_key = 544;
-  down_key = 545;
-  left_key = 546;
-  right_key = 547;
-  l1_key = 310;
-  l2_key = 312;
-  l3_key = 707;
-  r1_key = 311;
-  r2_key = 313;
-  r3_key = 708;
-  deadzone_y = 300;
-  deadzone_x = -300;
-  strcpy(inputstr, "/dev/input/by-path/platform-odroidgo3-joypad-event-joystick");
-}
-
-// ******************* OGX
-
-void handle_event_ogx(int type, int code, int value) {
-  if (type == 1) {
-    if (code == back_key && value == 1) {
-      emit(EV_KEY, back, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == back_key && value == 0) {
-      emit(EV_KEY, back, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == start_key && value == 1) {
-      if (hold == 1) {
-        system(quit_command);
-        system("sleep 0.2");
-        exit (0);
-      }
-      emit(EV_KEY, start, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == start_key && value == 0) {
-      emit(EV_KEY, start, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == a_key && value == 1) {
-      emit(EV_KEY, a, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == a_key && value == 0) {
-      emit(EV_KEY, a, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == b_key && value == 1) {
-      emit(EV_KEY, b, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == b_key && value == 0) {
-      emit(EV_KEY, b, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == x_key && value == 1) {
-      emit(EV_KEY, x, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == x_key && value == 0) {
-      emit(EV_KEY, x, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == y_key && value == 1) {
-      emit(EV_KEY, y, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == y_key && value == 0) {
-      emit(EV_KEY, y, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == l1_key && value == 1) {
-      emit(EV_KEY, l1, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == l1_key && value == 0) {
-      emit(EV_KEY, l1, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == l2_key && value == 1) {
-      emit(EV_KEY, l2, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == l2_key && value == 0) {
-      emit(EV_KEY, l2, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == l3_key && value == 1) {
-      hold = 1;
-      emit(EV_KEY, l3, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == l3_key && value == 0) {
-      hold = 0;
-      emit(EV_KEY, l3, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == r1_key && value == 1) {
-      emit(EV_KEY, r1, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == r1_key && value == 0) {
-      emit(EV_KEY, r1, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == r2_key && value == 1) {
-      emit(EV_KEY, r2, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == r2_key && value == 0) {
-      emit(EV_KEY, r2, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == r3_key && value == 1) {
-      hold = 1;
-      emit(EV_KEY, r3, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == r3_key && value == 0) {
-      hold = 0;
-      emit(EV_KEY, r3, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    //dpad
-    if (code == up_key && value == 1) {
-      emit(EV_KEY, up, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == up_key && value == 0) {
-      emit(EV_KEY, up, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == down_key && value == 1) {
-      emit(EV_KEY, down, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == down_key && value == 0) {
-      emit(EV_KEY, down, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == left_key && value == 1) {
-      emit(EV_KEY, left, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == left_key && value == 0) {
-      emit(EV_KEY, left, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == right_key && value == 1) {
-      emit(EV_KEY, right, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == right_key && value == 0) {
-      emit(EV_KEY, right, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-  }
-
-  // analog
-  if (type == 3) {
-    // mouse movement, right analog
-    if (right_analog_mouse) {
-      if (code == 4) { // up/down
-        if (value > deadzone_y) {
-          emit(EV_REL, REL_Y, -1);
-          emit(EV_SYN, SYN_REPORT, 0);
-        }
-
-        if (value < deadzone_x) {
-          emit(EV_REL, REL_Y, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-        }
-      }
-      else if (code == 3) { // left/right
-        if (value > deadzone_y) {
-          emit(EV_REL, REL_X, -1);
-          emit(EV_SYN, SYN_REPORT, 0);
-        }
-
-        if (value < deadzone_x) {
-          emit(EV_REL, REL_X, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-        }
-      }
-    }
-    else {
-      if (code == 4) { // up/down
-        if (value > deadzone_y) {
-          emit(EV_KEY, right_analog_down, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          right_analog_y = 2;
-        }
-        else {
-          if (right_analog_y > 0 || right_analog_x > 0) {
-            emit(EV_KEY, right_analog_down, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            right_analog_y = right_analog_y - 1;
-          }
-        }
-        if (value < deadzone_x) {
-          emit(EV_KEY, right_analog_up, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          right_analog_y = 2;
-        }
-        else {
-          if (right_analog_y > 0 || right_analog_x > 0) {
-            emit(EV_KEY, right_analog_up, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            right_analog_y = right_analog_y - 1;
-          }
-        }
-      }
-
-      if (code == 3) { // left/right
-        if (value > deadzone_y) {
-          emit(EV_KEY, right_analog_right, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          right_analog_x = 2;
-        }
-        else {
-          if (right_analog_x > 0 || right_analog_y > 0) {
-            emit(EV_KEY, right_analog_right, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            right_analog_x = right_analog_x - 1;
-          }
-        }
-        if (value < deadzone_x) {
-          emit(EV_KEY, right_analog_left, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          right_analog_x = 2;
-        }
-        else {
-          if (right_analog_x > 0 || right_analog_y > 0) {
-            emit(EV_KEY, right_analog_left, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            right_analog_x = right_analog_x - 1;
-          }
-        }
-      }
-    }
-    // mouse movement, left analog
-    if (left_analog_mouse) {
-      if (code == 1) { // up/down
-        if (value > deadzone_y) {
-          emit(EV_REL, REL_Y, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          left_analog_y = 2;
-        }
-        else {
-          if (left_analog_y > 0 || left_analog_x > 0) {
-            emit(EV_REL, REL_Y, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            left_analog_y = left_analog_y - 1;
-          }
-        }
-        if (value < deadzone_x) {
-          emit(EV_REL, REL_Y, -1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          left_analog_y = 2;
-        }
-        else {
-          if (left_analog_y > 0 || left_analog_x > 0) {
-            emit(EV_REL, REL_Y, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            left_analog_y = left_analog_y - 1;
-          }
-        }
-      }
-      if (code == 0) { // left analog x axis
-        if (value > deadzone_y) {
-          emit(EV_REL, REL_X, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          left_analog_x = 2;
-        }
-        else {
-          if (left_analog_x > 0 || left_analog_y > 0) {
-            emit(EV_REL, REL_X, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            left_analog_x = left_analog_x - 1;
-          }
-        }
-        if (value < deadzone_x) {
-          emit(EV_REL, REL_X, -1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          left_analog_x = 2;
-        }
-        else {
-          if (left_analog_x > 0 || left_analog_y > 0) {
-            emit(EV_REL, REL_X, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            left_analog_x = left_analog_x - 1;
-          }
-        }
-      }
-    }
-    else {
-      if (code == 1) { // left analog y axis
-        if (value > deadzone_y) {
-          emit(EV_KEY, left_analog_down, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          left_analog_y = 2;
-        }
-        else {
-          if (left_analog_y > 0 || left_analog_x > 0) {
-            emit(EV_KEY, left_analog_down, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            left_analog_y = left_analog_y - 1;
-          }
-        }
-        if (value < deadzone_x) {
-          emit(EV_KEY, left_analog_up, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          left_analog_y = 2;
-        }
-        else {
-          if (left_analog_y > 0 || left_analog_x > 0) {
-            emit(EV_KEY, left_analog_up, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            left_analog_y = left_analog_y - 1;
-          }
-        }
-      }
-      if (code == 0) { // left analog x axis
-        if (value > deadzone_y) {
-          emit(EV_KEY, left_analog_right, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          left_analog_x = 2;
-        }
-        else {
-          if (left_analog_x > 0 || left_analog_y > 0) {
-            emit(EV_KEY, left_analog_right, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            left_analog_x = left_analog_x - 1;
-          }
-        }
-        if (value < deadzone_x) {
-          emit(EV_KEY, left_analog_left, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          left_analog_x = 2;
-        }
-        else {
-          if (left_analog_x > 0 || left_analog_y > 0) {
-            emit(EV_KEY, left_analog_left, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            left_analog_x = left_analog_x - 1;
-          }
-        }
-      }
-    }
-  }
-}
-
-void config_ogx(char *inputstr)
-{
-#ifdef DEBUG
-  printf("OGA Contols - Configuring 'OGA/OGA 1.1' device\n");
-#endif
-  handleEventFunction = &handle_event_ogx;
-  back_key = 704;
-  start_key = 709;
-  a_key = 305;
-  b_key = 304;
-  x_key = 307;
-  y_key = 308;
-  up_key = 544;
-  down_key = 545;
-  left_key = 546;
-  right_key = 547;
-  l1_key = 310;
-  l2_key = 312;
-  l3_key = 705; /*minus key*/
-  r1_key = 311;
-  r2_key = 313;
-  r3_key = 708; /*plus key*/
-  deadzone_y = 300;
-  deadzone_x = -300;
-  strcpy(inputstr, "/dev/input/by-path/platform-odroidgo2-joypad-event-joystick");
-}
-
-// ******************* RK2020
-
-void handle_event_rk2020(int type, int code, int value) {
-  if (type == 1) {
-    if (code == back_key && value == 1) {
-      hold = 1;
-      emit(EV_KEY, back, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == back_key && value == 0) {
-      hold = 0;
-      emit(EV_KEY, back, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == start_key && value == 1) {
-      if (hold == 1) {
-        system(quit_command);
-        system("sleep 0.2");
-        exit (0);
-      }
-      emit(EV_KEY, start, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == start_key && value == 0) {
-      emit(EV_KEY, start, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == a_key && value == 1) {
-      emit(EV_KEY, a, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == a_key && value == 0) {
-      emit(EV_KEY, a, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == b_key && value == 1) {
-      emit(EV_KEY, b, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == b_key && value == 0) {
-      emit(EV_KEY, b, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == x_key && value == 1) {
-      emit(EV_KEY, x, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == x_key && value == 0) {
-      emit(EV_KEY, x, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == y_key && value == 1) {
-      emit(EV_KEY, y, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == y_key && value == 0) {
-      emit(EV_KEY, y, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == l1_key && value == 1) {
-      emit(EV_KEY, l1, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == l1_key && value == 0) {
-      emit(EV_KEY, l1, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == l2_key && value == 1) {
-      emit(EV_KEY, l2, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == l2_key && value == 0) {
-      emit(EV_KEY, l2, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == r1_key && value == 1) {
-      emit(EV_KEY, r1, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == r1_key && value == 0) {
-      emit(EV_KEY, r1, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == r2_key && value == 1) {
-      emit(EV_KEY, r2, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == r2_key && value == 0) {
-      emit(EV_KEY, r2, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    //dpad
-    if (code == up_key && value == 1) {
-      emit(EV_KEY, up, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == up_key && value == 0) {
-      emit(EV_KEY, up, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == down_key && value == 1) {
-      emit(EV_KEY, down, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == down_key && value == 0) {
-      emit(EV_KEY, down, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == left_key && value == 1) {
-      emit(EV_KEY, left, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == left_key && value == 0) {
-      emit(EV_KEY, left, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == right_key && value == 1) {
-      emit(EV_KEY, right, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == right_key && value == 0) {
-      emit(EV_KEY, right, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-  }
-
-  // analog
-  if (type == 3) {
-    // mouse movement, left analog
-    if (left_analog_mouse) {
-      if (code == 3) { // up/down
-        if (value > deadzone_y) {
-          emit(EV_REL, REL_Y, -1);
-          emit(EV_SYN, SYN_REPORT, 0);
-        }
-
-        if (value > 0 && value < deadzone_x) {
-          emit(EV_REL, REL_Y, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-        }
-      }
-      else if (code == 2) { // left/right
-        if (value > deadzone_y) {
-          emit(EV_REL, REL_X, -1);
-          emit(EV_SYN, SYN_REPORT, 0);
-        }
-
-        if (value > 0 && value < deadzone_x) {
-          emit(EV_REL, REL_X, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-        }
-      }
-    }
-    else {
-      if (code == 1) { // left analog y axis
-        if (value > deadzone_y) {
-          emit(EV_KEY, left_analog_down, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          left_analog_y = 2;
-        }
-        else {
-          if (left_analog_y > 0 || left_analog_x > 0) {
-            emit(EV_KEY, left_analog_down, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            left_analog_y = left_analog_y - 1;
-          }
-        }
-        if (value < deadzone_x) {
-          emit(EV_KEY, left_analog_up, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          left_analog_y = 2;
-        }
-        else {
-          if (left_analog_y > 0 || left_analog_x > 0) {
-            emit(EV_KEY, left_analog_up, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            left_analog_y = left_analog_y - 1;
-          }
-        }
-      }
-      if (code == 0) { // left analog x axis
-        if (value > deadzone_y) {
-          emit(EV_KEY, left_analog_right, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          left_analog_x = 2;
-        }
-        else {
-          if (left_analog_x > 0 || left_analog_y > 0) {
-            emit(EV_KEY, left_analog_right, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            left_analog_x = left_analog_x - 1;
-          }
-        }
-        if (value < deadzone_x) {
-          emit(EV_KEY, left_analog_left, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          left_analog_x = 2;
-        }
-        else {
-          if (left_analog_x > 0 || left_analog_y > 0) {
-            emit(EV_KEY, left_analog_left, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            left_analog_x = left_analog_x - 1;
-          }
-        }
-      }
-    }
-  }
-}
-
-void config_rk2020(char *inputstr)
-{
-#ifdef DEBUG
-  printf("OGA Contols - Configuring 'RK2020' device\n");
-#endif
-  handleEventFunction = &handle_event_rk2020;
-  back_key = 704;
-  start_key = 709;
-  a_key = 305;
-  b_key = 304;
-  x_key = 307;
-  y_key = 308;
-  up_key = 544;
-  down_key = 545;
-  left_key = 546;
-  right_key = 547;
-  l1_key = 310;
-  l2_key = 706;
-  r1_key = 311;
-  r2_key = 707;
-  deadzone_y = 300;
-  deadzone_x = -300;
-  strcpy(inputstr, "/dev/input/by-path/platform-odroidgo2-joypad-event-joystick");
-}
-
-// ******************* ANBERNIC
-
-void handle_event_anbernic(int type, int code, int value) {
-  if (type == 1) {
-    if (code == back_key && value == 1) {
-      hold = 0;
-
-      emit(EV_KEY, back, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == back_key && value == 0) {
-      hold = 0;
-      emit(EV_KEY, back, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == back_key && value == 2) {
-      hold = 1;
-    }
-
-    if (code == start_key && value == 1) {
-      if (hold == 1) {
-        system(quit_command);
-        system("sleep 0.2");
-        exit (0);
-      }
-
-      emit(EV_KEY, start, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == start_key && value == 0) {
-      emit(EV_KEY, start, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == a_key && (value == 1 || value == 2)) {
-      emit(EV_KEY, a, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == a_key && value == 0) {
-      emit(EV_KEY, a, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == b_key && (value == 1 || value == 2)) {
-      emit(EV_KEY, b, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == b_key && value == 0) {
-      emit(EV_KEY, b, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == x_key && (value == 1 || value == 2)) {
-      emit(EV_KEY, x, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == x_key && value == 0) {
-      emit(EV_KEY, x, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == y_key && (value == 1 || value == 2)) {
-      emit(EV_KEY, y, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == y_key && value == 0) {
-      emit(EV_KEY, y, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == l1_key && (value == 1 || value == 2)) {
-      emit(EV_KEY, l1, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == l1_key && value == 0) {
-      emit(EV_KEY, l1, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == l2_key && (value == 1 || value == 2)) {
-      emit(EV_KEY, l2, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == l2_key && value == 0) {
-      emit(EV_KEY, l2, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == l3_key && (value == 1 || value == 2)) {
-      emit(EV_KEY, l3, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == l3_key && value == 0) {
-      emit(EV_KEY, l3, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == r1_key && (value == 1 || value == 2)) {
-      emit(EV_KEY, r1, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == r1_key && value == 0) {
-      emit(EV_KEY, r1, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == r2_key && (value == 1 || value == 2)) {
-      emit(EV_KEY, r2, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == r2_key && value == 0) {
-      emit(EV_KEY, r2, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == r3_key && (value == 1 || value == 2)) {
-      emit(EV_KEY, r3, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == r3_key && value == 0) {
-      emit(EV_KEY, r3, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-  }
-
-  // d-pad
-  if (type == 3) {
-    if (code == up_key && value == -1) {
-        emit(EV_KEY, up, 1);
-        emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == up_key && value == 0) {
-      emit(EV_KEY, up, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    if (code == down_key && value == 1) {
-      emit(EV_KEY, down, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == down_key && value == 0) {
-      emit(EV_KEY, down, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == left_key && value == -1) {
-      emit(EV_KEY, left, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == left_key && value == 0) {
-      emit(EV_KEY, left, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-
-    if (code == right_key && value == 1) {
-      emit(EV_KEY, right, 1);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-    else if (code == right_key && value == 0) {
-      emit(EV_KEY, right, 0);
-      emit(EV_SYN, SYN_REPORT, 0);
-    }
-  }
-
-  // analog
-  if (type == 3) {
-    // mouse movement, left analog
-    if (right_analog_mouse) {
-      if (code == 5) { // up/down
-        if (value > deadzone_y) {
-          emit(EV_REL, REL_Y, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          left_analog_y = 2;
-        }
-        else {
-          if (left_analog_y > 0 || left_analog_x > 0) {
-            emit(EV_REL, REL_Y, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            left_analog_y = left_analog_y - 1;
-          }
-        }
-
-        if (value >= 0 && value < deadzone_x) {
-          emit(EV_REL, REL_Y, -1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          left_analog_y = 2;
-        }
-        else {
-          if (left_analog_y > 0 || left_analog_x > 0) {
-            emit(EV_REL, REL_Y, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            left_analog_y = left_analog_y - 1;
-          }
-        }
-      }
-      else if (code == 4) { // left/right
-        if (value > deadzone_y) {
-          emit(EV_REL, REL_X, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          left_analog_x = 2;
-        }
-        else {
-          if (left_analog_y > 0 || left_analog_x > 0) {
-            emit(EV_REL, REL_X, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            left_analog_x = left_analog_x - 1;
-          }
-        }
-
-        if (value >= 0 && value < deadzone_x) {
-          emit(EV_REL, REL_X, -1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          left_analog_x = 2;
-        }
-        else {
-          if (left_analog_y > 0 || left_analog_x > 0) {
-            emit(EV_REL, REL_X, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            left_analog_x = left_analog_x - 1;
-          }
-        }
-      }
-    }
-    else {
-      if (code == 5) { // up/down
-        if (value > deadzone_y) {
-          emit(EV_KEY, right_analog_down, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-        }
-        else {
-          emit(EV_KEY, right_analog_down, 0);
-          emit(EV_SYN, SYN_REPORT, 0);
-        }
-
-        if (value >= 0 && value < deadzone_x) {
-          emit(EV_KEY, right_analog_up, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-        }
-        else {
-          emit(EV_KEY, right_analog_up, 0);
-          emit(EV_SYN, SYN_REPORT, 0);
-        }
-      }
-
-      if (code == 4) { // left/right
-        if (value > deadzone_y) {
-          emit(EV_KEY, right_analog_right, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-        }
-        else {
-          emit(EV_KEY, right_analog_right, 0);
-          emit(EV_SYN, SYN_REPORT, 0);
-        }
-
-        if (value >= 0 && value < deadzone_x) {
-          emit(EV_KEY, right_analog_left, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-        }
-        else {
-          emit(EV_KEY, right_analog_left, 0);
-          emit(EV_SYN, SYN_REPORT, 0);
-        }
-      }
-    }
-
-    // mouse movement, left analog
-    if (left_analog_mouse) {
-      if (code == 3) { // up/down
-        if (value > deadzone_y) {
-          emit(EV_REL, REL_Y, -1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          left_analog_y = 2;
-        }
-        else {
-          if (left_analog_y > 0 || left_analog_x > 0) {
-            emit(EV_REL, REL_Y, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            left_analog_y = left_analog_y - 1;
-          }
-        }
-
-        if (value > 0 && value < deadzone_x) {
-          emit(EV_REL, REL_Y, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          left_analog_y = 2;
-        }
-        else {
-          if (left_analog_y > 0 || left_analog_x > 0) {
-            emit(EV_REL, REL_Y, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            left_analog_y = left_analog_y - 1;
-          }
-        }
-      }
-      else if (code == 2) { // left/right
-        if (value > deadzone_y) {
-          emit(EV_REL, REL_X, -1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          left_analog_x = 2;
-        }
-        else {
-          if (left_analog_y > 0 || left_analog_x > 0) {
-            emit(EV_REL, REL_X, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            left_analog_x = left_analog_x - 1;
-          }
-        }
-
-        if (value > 0 && value < deadzone_x) {
-          emit(EV_REL, REL_X, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-          left_analog_x = 2;
-        }
-        else {
-          if (left_analog_y > 0 || left_analog_x > 0) {
-            emit(EV_REL, REL_X, 0);
-            emit(EV_SYN, SYN_REPORT, 0);
-            left_analog_x = left_analog_x - 1;
-          }
-        }
-      }
-    }
-    else {
-      if (code == 3) { // w/s
-        if (value > deadzone_y) {
-          emit(EV_KEY, left_analog_up, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-        }
-        else {
-          emit(EV_KEY, left_analog_up, 0);
-          emit(EV_SYN, SYN_REPORT, 0);
-        }
-
-        if (value >= 0 && value < deadzone_x) {
-          emit(EV_KEY, left_analog_down, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-        }
-        else {
-          emit(EV_KEY, left_analog_down, 0);
-          emit(EV_SYN, SYN_REPORT, 0);
-        }
-      }
-
-      if (code == 2) { // a/d
-        if (value > deadzone_y) {
-          emit(EV_KEY, left_analog_left, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-        }
-        else {
-          emit(EV_KEY, left_analog_left, 0);
-          emit(EV_SYN, SYN_REPORT, 0);
-        }
-
-        if (value >= 0 && value < deadzone_x) {
-          emit(EV_KEY, left_analog_right, 1);
-          emit(EV_SYN, SYN_REPORT, 0);
-        }
-        else {
-          emit(EV_KEY, left_analog_right, 0);
-          emit(EV_SYN, SYN_REPORT, 0);
-        }
-      }
-    }
-  }
-}
-
-void config_anbernic(char *inputstr)
-{
-#ifdef DEBUG
-  printf("OGA Contols - Configuring 'Anbernic' device\n");
-#endif
-  handleEventFunction = &handle_event_anbernic;
-  back_key = 311;
-  start_key = 310;
-  a_key = 304;
-  b_key = 305;
-  x_key = 306;
-  y_key = 307;
-  up_key = 17;
-  down_key = 17;
-  left_key = 16;
-  right_key = 16;
-  l1_key = 308;
-  l2_key = 314;
-  l3_key = 312;
-  r1_key = 309;
-  r2_key = 315;
-  r3_key = 313;
-  deadzone_y = 2100;
-  deadzone_x = 1900;
-  strcpy(inputstr, "/dev/input/by-path/platform-ff300000.usb-usb-0:1.2:1.0-event-joystick");
-}
-
-// *******************
 
 // convert ASCII chars to key codes
 short char_to_keycode(char str[]) {
